Move additional code to LexiconPresenter
