+++ GreekReference/src/main/java/com/benlinskey/greekreference/presenters/LexiconPresenter.java
@@ -20,6 +20,10 @@ import android.content.ContentResolver;
 import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
+import android.util.Log;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.widget.ListView;
 import android.widget.Toast;
 
 import com.benlinskey.greekreference.R;
@@ -38,6 +42,63 @@ public class LexiconPresenter {
         mView = view;
         mContext = context;
         mResolver = mContext.getContentResolver();
+    }
+
+    public void onCreateOptionsMenu(Menu menu) {
+        setLexiconFavoriteIcon(menu);
+    }
+
+    public void onPrepareOptionsMenu(Menu menu) {
+        setLexiconFavoriteIcon(menu);
+    }
+
+    /**
+     * Returns {@code true} if the word with the specified lexicon ID is
+     * a member of the favorites list.
+     * @param lexiconId the lexicon ID to check
+     * @return {@code true} if the specified word is a member of the
+     *     favorites list, or {@code false} otherwise
+     */
+    private boolean isFavorite(int lexiconId) {
+        String[] columns = new String[] {AppDataContract.LexiconFavorites._ID};
+        String selection = AppDataContract.LexiconFavorites.COLUMN_NAME_LEXICON_ID + " = ?";
+        String[] selectionArgs = new String[] {Integer.toString(lexiconId)};
+        Cursor cursor = mResolver.query(AppDataContract.LexiconFavorites.CONTENT_URI,
+                                        columns, selection, selectionArgs, null);
+
+        if (cursor == null) {
+            throw new NullPointerException("ContentResolver#query() returned null");
+        }
+
+        boolean result = false;
+        if (cursor.getCount() > 0) {
+            result = true;
+        }
+        cursor.close();
+        return result;
+    }
+
+    /**
+     * Sets the Lexicon Favorite icon to the appropriate state based on the currently selected
+     * lexicon entry.
+     * @param menu the {@code Menu} containing the Favorite icon
+     */
+    private void setLexiconFavoriteIcon(Menu menu) {
+        MenuItem addFavorite = menu.findItem(R.id.action_add_favorite);
+        MenuItem removeFavorite = menu.findItem(R.id.action_remove_favorite);
+
+        int id = mView.getSelectedLexiconId();
+        if (!mView.isDetailFragmentVisible() || id == ListView.NO_ID) {
+            // Hide both icons when the app is in one-pane mode or no item is selected.
+            addFavorite.setVisible(false);
+            removeFavorite.setVisible(false);
+        } else if (isFavorite(id)) {
+            addFavorite.setVisible(false);
+            removeFavorite.setVisible(true);
+        } else {
+            addFavorite.setVisible(true);
+            removeFavorite.setVisible(false);
+        }
     }
 
     public void onAddFavorite() {
@@ -86,6 +147,13 @@ public class LexiconPresenter {
         values.put(AppDataContract.LexiconHistory.COLUMN_NAME_LEXICON_ID, id);
         values.put(AppDataContract.LexiconHistory.COLUMN_NAME_WORD, word);
         mResolver.insert(LexiconHistoryProvider.CONTENT_URI, values);
+    }
+
+    public void onClearFavorites() {
+        mResolver.delete(AppDataContract.LexiconFavorites.CONTENT_URI, null, null);
+
+        String msg = mContext.getString(R.string.toast_clear_lexicon_favorites);
+        mView.displayToast(msg);
     }
 
     /**
