+++ GreekReference/src/main/java/com/benlinskey/greekreference/views/lexicon/LexiconDetailView.java
@@ -26,4 +26,6 @@ public interface LexiconDetailView {
     void invalidateOptionsMenu();
 
     void displayToast(String msg);
+
+    boolean isDetailFragmentVisible();
 }
